deleted moving py files to folder, did it in other branch
