+++ check_registration_with_bugz.py
