+++ get_attribute.py
